ServerSentEvents/field-name-other.html

<!DOCTYPE html>
<html>
<head>
	<title>The steps to process the field given a field name and a field value depend on the field name: Other:</title>
  <meta charset="utf-8" />
  	<meta name=viewport content="width=device-width, maximum-scale=1.0, user-scalable=no" />
	  <link rel="author" title="Fabrice Clari" href="mailto:f.clari@inno-group.com">
	  <link rel="author" title="Dimitri Bocquet" href="mailto:Dimitri.Bocquet@mosquito-fp7.eu">
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	
	
</head>
<body>
	<h1>The steps to process the field given a field name and a field value depend on the field name: Other:</h1>
	<div id="log"></div>
  <script>
  	setup({timeout:6000});
    var s = new EventSource('http://html5.mosquito-fp7.eu/messages/t0078');
    var t = async_test("The field name is other: the field is not 'data'.");
    var u = async_test("The field name is other: the field is not 'id'.");
 
	
    s.onmessage = function(e){
    		var dat = e.data.split("\n");
    		t.step(function(){assert_not_equals(dat[1], "other");});
    		u.step(function(){assert_not_equals(e.lastEventId, "other");});
    		t.done();
    		u.done();
    };
    


    

  </script>
</body>
</html>
